refactor: Move SettingsProvider to context and update imports
